Refactor, move validator CLI related stuff to its own module(cli.rs)
The main function of the validator is getting big. Improve readability by moving away the CLI setup stuff to its own module.
